+++ autogen/agentchat/__init__.py
@@ -2,7 +2,6 @@ from .agent import Agent
 from .assistant_agent import AssistantAgent
 from .conversable_agent import ConversableAgent
 from .groupchat import GroupChat, GroupChatManager
-from .multimodal_conversable_agent import MultimodalConversableAgent
 from .user_proxy_agent import UserProxyAgent
 
 __all__ = [
@@ -12,5 +11,4 @@ __all__ = [
     "UserProxyAgent",
     "GroupChat",
     "GroupChatManager",
-    "MultimodalConversableAgent",
 ]
