+++ test/agentchat/contrib/test_lmm.py
@@ -1,7 +1,14 @@
+import autogen
+from autogen.agentchat.agent import Agent
+from autogen.agentchat.contrib.multimodal_conversable_agent import MultimodalConversableAgent
+
+import os
+import sys
 import unittest
 from unittest.mock import MagicMock
 
-from autogen.agentchat import Agent, MultimodalConversableAgent
+KEY_LOC = "notebook"
+OAI_CONFIG_LIST = "OAI_CONFIG_LIST"
 
 base64_encoded_image = (
     "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4"
@@ -11,7 +18,21 @@ base64_encoded_image = (
 
 class TestMultimodalConversableAgent(unittest.TestCase):
     def setUp(self):
-        self.agent = MultimodalConversableAgent(name="TestAgent")
+        config_list = autogen.config_list_from_json(
+            OAI_CONFIG_LIST,
+            file_location=KEY_LOC,
+            filter_dict={
+                "model": ["gpt-4", "gpt4", "gpt-4-32k", "gpt-4-32k-0314", "gpt-4-32k-v0314"],
+            },
+        )
+        self.agent = MultimodalConversableAgent(
+            name="TestAgent",
+            llm_config={
+                "timeout": 600,
+                "seed": 42,
+                "config_list": config_list,
+            },
+        )
 
     def test_system_message(self):
         # Test default system message
